Stack, tiny interface, parsing tags
Setup a stack and got a tiny interface going. The interface also
supports the E9 hack if it the makefile is invoked with e9hack as one of
the targets. I do have one concern, which is that the data pushed onto
the stack when calling outb is not pulled off the stack until the caller
function returns. As for parsing the tags structure that multiboot2
gives me: something is wrong, very wrong. There seems to be 880 bytes
worth of tags, and the first tag doesn't have a valid uint32_t type.