More information printing
Just made a few more things, added some more features to the interface.
Such as being able to change the number of hex characters a hex
characters printed to the screen. Also gave the module, the kernel, a
name which the bootstrapper will use to recognize it:
"arctan-module.kernel.elf".