Depth issue fixed and scrolling
I noticed that one character was being left out when printing a number
and having a depth which is > 0, this was fixed by changing the initial
value of r to be 0 instead of 1. Implemented simple memcpy and made the
screen scroll. The cursor also now moves because why not. Please excuse
the big block of prints in the main function. It was definitely just
for testing the scroll functionality (definitely no silly business here)
.