Getting to the kernel
Working on parsing out the Kernel ELF module so that the bootstrapper
can jump into the kernel. For this a little ELF parser will be used.
The framebuffer request tag was removed from the multiboot header for
testing purposes, for when the kernel is successfully called a 'B' will
be displayed as the first character.