Framebuffer and information
Working on the framebuffer system, separating it out into its own field
of C files. Also working on parsing the information structure given by
GRUB. Currently, initialization of the framebuffer is not working, as
there is no white dot placed onto the screen, but the general ground
works are there. I'll make it possible to layer mutliple framebuffers
together. This way, for the terminal, there can be a graphics buffer and
a text buffer. Then there will be an ultimate framebuffer which will mix
all of these sub-buffers together in a given order.