Memory management
Finished with the page frame allocator, currently trying to wrap my
head around how I should get to work with the virtual memory manager
and an allocator. Started work on the virtual memory manager. Looking
at some other operating systems, vmm.c is typically responsible for
managing the page tables. So far, my vmm.c contains a stack of memory
ranges. It works on the premise that the first element is located in the
middle, and there are two hemispheres. Allocations alternate from the
positive hemisphere to the negative hemisphere. This would hold the
structure of virtual memory, allowing for mapping to be easier, but I'm
at an impass, my method or a proven method.