boot.asm fixes and allocator start
Fixed a few mistakes in boot.asm. These mistakes were related to the
boot header, which is not something that should have mistakes in it.
Started working on the allocator so that there is a simple way of
managing page frames in the kernel heap. I will refactor the PMM so that
it uses the allocator, and this will let me tie in VMM no problem.