Free list allocator
Saw a code snippet of a free list allocator, it clicked in my mind
instantaniously. Changing the simple watermark allocator
to a free list allocator. I might make it more generalized so that it
can be used a general pool allocator. Which it currently is for the
kernel. More generalized functions would look like:
alloc(pool, page_count); or free(pool, address, page_count);. Currently
allocating more than page is implemented, going to work on freeing more
than one page and testing the entire system, so far it is doing a good
job of replacing the watermark allocator.