Allocator troubles
While the allocator in the bootstrapper may not be causing all of the
problems I am facing currently - it is a contributor. The addresses are
all wonky, which I do not appreciate. When we first allocate we get an
address of 0x10B000 which is good. When we allocate for a second time we
get an address of 0x90B000, which is not good, we would expect an
address of 0x10C000. After that it would be best to run over the logic
behind setting up the HHDM.