Ruled out printf and allocator
I have hooked up a pre-existing printf implementation in the interface.
It works like a charm, and my printfs have been shortened a great deal,
although I had to add "arith64.h". This is because there are some
runtime functions that I simply cannot get hold of. Corrected the
allocator, regions of available memory are now connected and the
allocator properly allocates a single page. Everything still triple
faults, but that is probably do to how the page tables are constructed.