Got to the kernel
I needed to disable SSE2 as well if I wanted the printfs to work. I got
the mappings working, the kernel was replacing the HHDM in the PML4 due
to the offsets being calculated from the kernel's offset rather than the
kernel's vaddr. On that note: got rid of the kernel_info table and
replaced it with two uint64_ts - kernel_vaddr and kernel_elf_off.