Start of the junction
Worked on getting information from the interrupt stub over to the
junction, it has went well. We now catch interrupts in 32-bit mode. Now
I need to work on parsing out the information within the junction and
calling different more specific handlers based on the interrupt that was
received, for now we just print out the interrupt that was called (code
and name), registers, and change EAX to be something we can spot.