Narrowed kernel problem
I got the kernel somewhat operational. The segmentation faults and the
misalignment seems to be because of some combination of compiler flags
and code. Removing code that uses the pool allocator and the
initialization of the PMM, everything can be drawn fine. This tells me
that my issues are somewhere in the PMM and in the pool allocator.