Created generalized freelist allocator
I created a generalized, simple freelist allocator for the kernel. This
will be used in the PMM. The next step is to get a 64-bit IDT going so
we can get some more detailed information as to why we triple fault.
Currently the system just does not want to write to the framebuffer for
whatever reason, nor do the printfs want to behave properly if I do not
have the `const char *string = "Hello World";` located within the code.