Kernel swapping idea
Wouldn't it be cool for the kernel to be able to swap itself out without
having to reboot? It could just say, new kernel here to the
bootstrapper, bootstrapper loads it, bootstrapper calls it. As long as
the two kernels are compatible one could pickup right where the other
left off.