arctan.h and correcting HHDM
The HHDM is supposed to be an identity map of the entirety of physical
memory located at the first higher-half address (0xFFFFFFFF00000000).
Worked on correcting my HHDM to be mapped there too, but I am
encountering a triple fault because the bootstrapper is not mapped into
memory. Created an arctan header file, this way I can have definitions
across both the kernel and the bootstrapper.