Fixed misalignment, SSE
The reason that the kernel would not work without the const char *string
= "Hello World"; is because ALNUM in temp/interface.c was not being
placed into the .rodata section. I manually told it using an attribute
to put into .rodata, and the problem immediately went away. Currently
working on porting the printf implementation from the bootloader over to
the kernel. It does involve some SSE stuff, which I am currently
implementing support for.