Memory, structures, and SSE
I added a new feature to the freelist.c, this is to combine two newly
created freelists into one. I also changed the kernel / bootloader
interactions to use the ARC_BootMeta structure. I worked on SSE,
however, it seems that not everything wants to play nice. If I do not
have SSE and AVX (for SIMD instructions), I will triple fault when
trying to initialize the PMM (because of some division requires movaps),
and if I do try to enable it then the CPU just stops when setting
CR4 (when CR4.OSXSAVE = 1).