Kernel rework
Got rid of basically the entire kernel and started from scratch. This is
so that it would be a little bit simpler to understand and debug
things. This is a working copy. Apparently adding to the program pointer
thing (. in the linker script) merely changes the address within the
symbol table. The data in the file is not offset. When I do not do that
it works. Something. Uh, well anyway, tomorrow is mostly going to be
dedicated to getting the kernel back to whatever it was before, but
working.