Kernel refactoring, printf problems
I started on the kernel rework, the directory structure is much cleaner.
However, in both the bootstrapper and kernel the printf implementation
is causing errors. These occur during runtime, and it seems like
everything just gets misaligned. For example once `printf("%"PRIX64"\n",
value);` had an output of "GO1))))". In the bootstrapper, the
capabilities of the ELF load are going to be polished up.