Bootstrap: VMM and checks
I wrote a little VMM that is able to just simply map a physical address
to a virtual address without being a logical pain. Got to work on some
of the checks. It is currently very strict, and assumes what the kernel
wants, this is something I need to revise at a later point.