Bootstrap: ELF parser and 64-bit entry
Created a simple ELF parser for the bootstrapper so that it is able to
map the kernel into memory. Also started working on the steps to get to
a 64-bit environment; however, I am experiencing some type of
fault (probably a #GP) when trying to long jump. CR2 is also set to be
the address of the IDTR + 0x80, which would be the 16th entry, I'm not
entirely sure why an FPU exception would be triggered by a long jump,
but it is worth noting.