Bootstrap: ELF mapper error
I have "found" why there is a fault when the system preforms a long jump
to 64-bit mode. For some odd reason the call to map_page in the ELF
parser messes up the page tables. I do not know why this is happening,
but it is happening only when the ELF parser maps pages into memory.