Kernel: called
The kernel can now be called and code is successfully executed from it.
Ported over the debug definitions from the bootstrapper to the kernel.
However, we are still facing some issues with the IDT in the kernel.
When anything besides an int0 is called, it double faults, meaning it
isn't able to handle it, at least not when I call it. Page faults are
handled / reported just fine.