Kernel: working IDT, broken GDT
Got the IDT working, the tables needed to be tagged with the attribute
__attribute__((section(".data"))), for some reason these weren't already
in the data section, but there is no difference made to the sizes of the
program headers. I am not sure why it works, but it does. When
preofrming the long jump in the kernel's GDT, the processor triple
faults for currently unknown reasons.