Screen printing
In the _putchar function I added some code which will draw a character
to the screen. Currently it isn't completely air tight, but it does
properly place characters on the screen. Further, the kernel is very
finicky, because for basically every variable it requires me to
manually add `__attribute__((section(".data")));`. If I do not do that,
the variable (initialized or not) will be set to random data.