Kernel: simplified boostrap, minimal keyboard
Simplified the bootstrapper. It does not really need to be display
anything to the screen. If we really want the big list of logs, we can
ask the kernel to the just iterate through the multiboot list again. I
implemented a minimal PS/2 keyboard handler as well just to see some
user input on the terminal. This is temporary.