Kernel: added font, Bootstrap: ELF errors
Added the kernel font, it was previously a .bin which was in the
.gitignore. The issues related to the
`__attribute__((section(".data")));` was occurring because the ELF
loader did not map the ELF file into memory properly. It mapped the
program headers instead of the actual sections. In the ELF
specification, the .bss section is not present in the file, and must be
allocated by the loader.