Kernel: successfully regains bootstrap PMM context
Changed freelist.h so that struct ARC_FreelistMeta and ARC_FreelistNode
have their members aligned to 8 bytes, this will make the 64-bits in size no
matter if compiled for 32-bit or 64-bit systems. There may be a better
way to acheive this, but this works for now. The kernel's PMM
successfully changes the addresses in the freelist to be in the HHDM.
One more thing to be done is to analyze the memory map given by GRUB,
and look for entries which are not yet incorporated into the PMM's
freelist.