Kernel: tested contiguous functions
I made a tiny test in kernel.c for the freelist's contiguous alloc and
free functions. They work, I have not tested if the fails system
works - where we put allocations into a new freelist and then free them
once we found a contiguous section. The order in which the contiguous
free free's addresses needed to be inverted.