Kernel: working vmm
The VMM, at least the mapping part of it, is in a function state. As
well as the now, fairly simple, invlpg. Made numerous silly mistakes
because of my macros and my code not obeying the flags structure. There
was also an issue where the PMM would some reason start allocating in
the first SLAB list, this was fixed by making the SLAB allocator
allocate memory for its lists using the PMM's wrapper.