Kernel: resources and references
Created references and implemented them. The resource.c file provides
functions to create and destroy references. References are going to have
functions which will describe things that the resource should do. Driver
functions are exported into sections, this allows for a init function to
not be used. The next step will be to refactor the initramfs driver so
that it fits in with the current driver model.