Kernel: initramfs driver, VFS, and resources
Worked a little bit on the initramfs driver. I was mostly adding some
new features to the VFS which needed to be implemented in the initramfs
driver. The open function should now properly clean itself up if it
fails to create the node graph. More clean up code needs to be added.
Resources now use only heap memory, therefore all resources can be
uninitialized. Driver definitions are now also aligned to 1 byte as to
not have padding between the various driver definitions.