Kernel: function name changes, VFS functions
I changed the name of some old functions so that they follow the current
naming convention. The link function now uses the create function in
order to create the links. The new link file copies the permissions
needed to access the source file. I need to figure out how links should
be saved to a filesystem in case that they're made under one, and how to
modify links.