VFS first tests
Implemented the create and the mount functions, currently working on the
unmount function. However, when I printed the error codes of
Arc_CreateVFS and Arc_MountVFS, they were both 0, indicating that
vfs_traverse is working as intended. I have also implemented strndup, as
well as a create_level in the vfs_traverse_info structure.