VFS success: open, seek, read, mount
The functions are working, though I want to go back through them and
thoroughly double check them. In vfs_traverse I also need to implement a
way to progressively build the statted path to get the actual stat for
each node  instead of just using info->mountpath, which will give the
stat for the file always.