VFS: fixed links
There was a weird bug with links, where the source file would not be
opened, and it would cause errors down the line in terms of no font
displaying or various faults. I have now fixed this by making links copy
the is_open attribute of their source file. If a link is detected to
be empty, its source file is marked open, and it is marked open.
There are now also place-holder resources, which serve to only hold
references. This is probably and inefficient implementation, but it
should work for now.