Resources: fundemental changes
I have changed how the resources, VFS, and drivers are going to handle
things. File descriptors are now defined in resource.h instead of vfs.h,
and driver functions use them for function calls. This does mess
somethings up regarding link resolution (i.e. where is the link
going to be resolved, because resolving it both in the vfs and the
driver is a bit wasteful). Queue locks can now be frozen and thawed.
When a queue is frozen, no more threads can lock it, and all other
threads will be completed before the locking thread. This serves to
eliminate the issue of a vfs node being used after it has been
deleted as this puts the signalling in place to stop that.