VFS: working fundemental changes
The fundemental changes described in the aforementioned commit are now
working. The resource pointer is now passed along to the driver
functions as well as the file descriptor pointer. A prune function
has also been implemented so that unused nodes are removed from the
VFS. This function does grab quite a big lock, so it may slow down in a
multiprocessing system.