ACPI: simple implementation
Wrote a simple ACPI implementation to just parse the tables to get to
the MADT, which contains valuable information about the IO APICs. In
apic.c there is also a public function which allows for a IO APIC
definition to be inserted into a inverted linked list. There is also a
bug in strncmp and probably strcmp as well.