APIC: driver / structure rework
I reworked the job of the structure, now the initialization function
creates a driver state that is 16-bytes in size. It is a linked list
containing pointers to the beginning of each entry within the data
section of the MADT. This list can be traversed using the read headless
read function. All work done to control the APICs will be done under
arch/x86-64/apic, with drivers/data/acpi/madt.c serving only to retrieve
information.