ACPI: FADT, DSDT, HPET, MADT checksum fix
Implementing the structure for the FADT, DSDT, and HPET tables. I have
also fixed a bug when the MADT driver goes to check its checksum.
Previously, it was using the size of the apic struct, rather than the
size of the table, thus giving an improper sum, and making the driver
think that the table was invalid.