cAML: parsing one package
The parser is now able to parse, on a surface level, one package out of
many. By surface level, I mean that it just hops over the various
definitions / ocpodes without actually building out the ACPI namespace.
However, it shouldn't be too tricky to implement the actual meat and
potatoes of the parsing.