Drivers: cleaning, new concepts
I cleaned up the initramfs code, so now the initramfs super driver code
is separated from the file driver code. The VFS now also checks to see
that a filepath has a length greater than 0. New concepts: buffers and
fifos in the VFS are ideally going to make IPC easier, as well as clean
up the weirdness with the ACPI namespace stuff.