VFS: hacks
Made something hacky. I needed to pass an argument through to the
creation of a resource in the vfs_traverse function, so I just made a
new field in the vfs_traverse structure which would hold this argument,
it feels awfully hacky. This is included alongside some changes to
accomdate FIFOs and buffers. I still need to write the drivers for
these.