cAML: work on EXTOPs and fixes
Fixed an issue when parsing TermLists, state->max would be lost,
resulting in the parsing being aborted. This was fixed by just properly
calculating _max, and setting the state's max value to this newly
calculated value. The CHECK_STATE macro would also automatically return
-1 upon the detection of an invalid state. I have made the action taken
customizable, this fixes the state not being restored when parsing a
TermList. Implemented parsing for some EXTOPs, though the information is
not made use of. I need to work out how exactly I will store the
information. The information also is mostly from the fixed arguments, or
the header of each operation, and dynamic arguments are simply gleamed
over. The NameString, for NULL Names, now returns a dynamically
allocated empty buffer instead of a static empty buffer, this is so that
the caller does not have to implement additional logic to know to not
free the static address.