Working order
The change in hte vfs_idx2idx function resulted in improper indices
being returned for drivers. This was because of a NULL check. Some
memory issues were also present, as some addresses were not being converted
from physical to HHDM, this resulted in some cases the addresses
slipping out of the first 4 MiB that is identity mapped by the MB2 bootstrapper.