PMM: logical fixes
The PMM had a logical issue with the while loop which converts all
physical addresses in the page frame freelist to HHDM addresses.
The issue was that the loop would not stop and end up at
ARC_HHDM_VADDR, and continue on until it would encounter a page fault.
This is because the check current->next != NULL would in some cases
fail as the 32-bit pointer may be NULL, the 64-bit pointer may have its
higher 32-bits filled with some random data. To solve this, the check
has been replaced with (uintptr_t)current != ARC_HHDM_VADDR. This
simplifies the loop greatly. One more thing to investigate is this
"random" data.