MB2 Rewrite: working on real hardware
After lot's of coercing, the kernel is finally back to where it was some
time ago - working on real hardware. Most of the issues were in the PMM,
which is expected because the rewrite of the bootstrapper changed it
ever so slightly to make all the old code look fine when it really
wasn't. Along the way some bits had to be cut off, like using a custom
font file, parsing ACPI, and APIC stuff. This prompts me to re-write or
tidy up the code a little.