VFS: not sure what I did but it works
Mount operations weren't creating the proper mount structure, it
probably should be omitted, or refactored heavily in some way. Added in
a check to make sure that mountpoints were mountpoints and not
devices (should not need to be checked). With these changes the VFS now
suddenly works on real hardware - very surprising considering how janky
some of the older code is in it.