VFS: minor refactor
Mount structures have now been removed. The number of open files and the
type of the filesystem do not need to be explicitly stored, as the
latter can be inferred from the driver group/index, the former isn't
that useful and can be kept track of in a different way. ACPI stuff has
been updated, there is no more separate directory for rsdt/xsdt rather
the rsdt and xsdt drivers are mounted on /dev/acpi. The buffer driver
also has a brand new check which checks if the driver state is NULL. The
VFS listing also displays the type of each node in human readable text.