Freelist: checked over
I checked over the freelist once more. There were a few problems which
are now fixed. Firstly: spelling of ceil. Second: when initializing a
freelist, there were missing checks regarding if the list could fit the
meta and at least one other object, the subtraction of objects from the
number of free objects in the list, and the ceil value being initialized
as a base of an entry (which could cause initialization of an entry
outside of the given memory range).