Refactoring Memory Management
I have not made a commit in quite a while, but all I have been doing is
just reworking the memory management structure. Currently I have
implemented an untested buddy allocator, created a pager (which is
architecture specific) and changed the top level function the kernel
calls to allocate memory. Currently allocations are broken.

In more detail:

The new architecture of memory management looks something like this:

   PMM--------Pager--------VMM--------Top Allocator
    |                       |               |
    |                     Buddy            Slab
    |                       |               |
 Freelist----------------Freelist--------Freelist

The PMM handles physical pages in the HHDM using a freelist. This is the basis of all
other allocators. To its right, the pager. This is the file which
manages the page tables, or however a specific architecture implements
virtual memory. All it needs to provide is a way to map, unmap, and set
attributes of virtual memory regions. With this, the VMM is now
possible, which uses a buddy allocator to manage the virtual address
space. The buddy allocator implementation itself uses a freelist to
manage the tree of objects it uses to keep track of the various
allocations. The top-most allocator, which is just a wrapper around a
Slab allocator, requests memory from the VMM, which allows it to
implement Arc_Alloc and Arc_Free.

Also, the following files are abstract implementations:
 - mm/buddy.c
 - mm/freelist.c
 - mm/slab.c

This should allow them to be used on any range of memory.