Buddy: working with the VMM test
There were a few corrections to be made to the buddy allocator
implementation. The quick_log2 function miscalculated the exponent, the size of the
first node was erroneously set. This has been fixed by it being able to
now return the estimated size, and by keeping track of the original size
to apply the log2 to. The various split function is now recursive and
executed in the initialization function, this should increase allocation
time. Freeing still needs to be tested.