VMM: rough implementation, Pager: attributes
I worked out a rough implementation for the VMM, I had to change the
buddy allocator so that when it would free an address it would return
the size instead of the address. This is so that the VMM can just
directly feed the return value into the unmap function of the pager.

Implemented architecture specific stuff in the pager to set the
attribute bits of the various page tables. Going to work out the unmap
and set attributes functions now.