Pager: my head hurts, corrected indentation
Been working on the pager, and my head is hurting from it. Currently the
regular and fly mappers work, but the unmap function seems to be having
a little bit of trouble. The attribute setting function still needs to
be created.

And just some general corrections regarding the indentation on switch
statements.