APIC: rough, but working
I have been messing around with the APIC just a little bit, I had most
of the things to do it ready, but I just needed to mess around with the
registers a little bit. However, now it works. I am just making this
commit so that I have a reference point of what actually works.