Atomics: added "mutexes" and spinlocks
Currently mutexes behave exactly like spinlocks, as I do not have a
scheduler which I can ask to yield the CPU and or wake up the current
thread. The thread ID system is also going to need a overhaul to also
encode the CPU the thread is on. With the mutexes and spinlocks
implemented I have encountered a few deadlocks in the freelist and in
the virtual filesystem. The silly lock in printf.c have now been
replaced with proper mutexes, and the lock moved to the terminal
structure. Finally I have another computer I can test Arctan on, a Dell
Optiplex 3020, which the kernel triple faults on when trying to do
draw_term or vfs_create. I will focus my priorities on draw_term as that
will provide me with information I need to debug vfs_create.