SMP: processor syncing, and misc stuff
I went in and got the various processors in the system all synced up. In
apic.c there is also no more linked list to keep track of all of the
LAPICs, rather they are embedded in the ARC_ProcessorDescriptor. As for
other things, I changed around how debug statements work, and that's
about it.