SMP: fixed some mistakes
LAPIC ID's aren't necessarily consecutive, so I've reworked the listing
function to work with this concept in mind. I have also moved some of
the initialization to move the AP to high memory into smp.asm. The
things that have been moved are: initialization of the higher half
stack, enabling of the XD bit, and loading the PAT. I will now begin
work on the mechanism that will allow me to redirect a given processor.