LAPIC timer
Created functions to manage the LAPIC timer. It is not yet calibrated,
but it does provide a nice timer interrupt. I have set up all processors
to now use the LAPIC timer and made sure that the PIT IRQ remains
masked. I may change this up so that interrupt 32 is the global timer
and the LAPIC timer uses some another interrupt line.