Architectural abstraction
Going to rework the current structure of the kernel so that the
architecture specific stuff is abstracted behind arch/. This commit
includes the first step of this which is to move mp/smp.c and mp/smp.h
into arch/x86-64 and to move arch/x86-64/acpi to arch/x86-64. This is
because smp.c and smp.h are specific to x86-64, while ACPI is
architecture independent. The next step will be to create a layer in
arch/ that can be called upon by the rest of the kernel. This layer will
properly resolve to the target architecture at compile time.