Atomics: tickets, VFS: concurrency nightmares
Renamed and slightly changed the functionality of queue locks. They are
now ticket locks, and use unique ticket IDs instead of only using thread
IDs.

The VFS is causing some problems regarding concurrency. See proc_test in
kernel main. This function opens one file in the initramfs, reads it,
closes it. However, when done across multiple cores some problems are
encountered. For instance, if all cores finishing traversing the VFS and
lock the credit.txt node before a close is called, everything works just
fine. On the other hand, if one core is ahead of the others, and decides
to close the file before the others managed to hand the lock down in the
vfs_traverse function, the close function deadlocks.