VFS: fixed concurrency nightmare
The vfs_traverse function assured the caller "yes yes, I incremented the
ref_count for you and you own the branch_lock when I return". This
partially true, but worst of all: partially a lie. As it, vfs_traverse,
never actually yielded the ticket lock on the final node found, rather
it just returned. Anyway, that's fixed now, going to tidy up the code
later if it needs it, and I simply cannot wait for the next concurrency
bug!!!!! YAY!!!!!