VFS: SMP create and write tests
Doing some more testing with SMP on the VFS. The proc_test function now
also creates and writes a buffer, that is then read back by the BSP
successfully. There are some bugs in the VFS, for instance, when it says
it is closing and deleting the credit.txt node it doesn't, so there are
some things that I need to sort out.