VFS: changed seeking, links work
Seeking code has been moved from each individual driver to the
higher-level vfs_seek function. Each driver's seek function now only
aims to inform the driver that the read-write has been moved and to
update accordingly. Links have been fixed and now work. This needed a
new function in the mount driver definition called "locate". This
function serves to gather information and create a structure which is
passed to a file driver's initialization function as the args parameter.
This allows for the remove of filepath from the open function, as it is
no longer needed for opening the file.