VFS: permissions and error checking
Implemented proper permissions checking into the VFS, as well as scanned
over all fail cases. The error cases all look to be fine (I also went
in and added some of the missing cases / simplified older ones).
However, I still do not trust it the error handling fully, the VFS is
one of the most complex mechanisms in the kernel.