MM: small buddy fixes, larger allocator.c allocs
There were some minor issues in the buddy allocator, which prevented it
from functioning with multiple allocations (the point of an allocator).
Made it possible for the kernel allocator to allocate more than 2048
bytes at a time by utilizing the VMM. The pager and SLAB allocator do
complain as currently, the latter because free(a); pokes it to see if
the given address is part of the SLAB allocator or not. The pager, I am
not quite sure why; it definitely needs a rework.