MM: implemented rest of pager, SLAB silencing
Implemented the rest of the pager functions. I also implemented an
attribute for the SLAB allocator to disable the error message on the
slab_free function. This is because mm/allocator.c blindly calls
slab_free on its free function, and if the allocation is a larger one,
this causes an error to be reported. This is for log cleanliness, as in
this case, it is not an error. The free function will now only report an
error given that both the slab_free and vmm_free functions return NULL.