cAML2: iterative design, compiling time
The iterative design is working. There is no need for scopes or a scope
stack, as that can be handled when getting a name. This should be
refined so that the get_name_string function is not responsbile for
changing the current directory. I am thinking about using filepaths
instead of node pointers, since that may be a little bit nicer and would
not require a vfs_create_rel_node function. Lastly, I am beginning work
on the compiler to translate AML into x86-64 machine code.