Big explosion
The kernel has now exploded into many different repositories. This is so
the various modules constituting the kernel can be individually tested,
as well as changes to them more specifically defined in commit messages.
This may also help minimize problems if portions of the kernel were to
be re-written in different languages, theoretically all it requires is:
the language is able to properly define an interface that may be used by
the rest of the kernel, define the proper symbols in generated object
files, and, of course, generate the proper ELF object files.