Added delete function to the VFS
A concern is that of deadlocks. In the event that a delete function
attains the ticket lock, it does not freeze it - this is because errors
were encountered with the ticket lock function returning NULL in the
event that a ticket lock was frozen. However, not freezing opens the
possibility of destroying the ticket lock and halting processes. This
problem would persist with mutexes. In the current circumstances, this
should not happen; however, if scaled up it could become a serious
issue.