Partitions and ordered build system
Partitions can now be detected, so I've just implemented a simple little
read test. Lastly, the build system was giving me a lot of trouble, and
I will need to rethink this fairly hacky solution so that it is a little
bit nicer; however, now ordering is enforced such that the dri_defs are
produced before the code is compiled, and the code is compiled before it
is linked.