Update modules, clean kernel before build
Updated the modules to the latest and I made it so that the kernel is
cleaned of object files before it is built. The old multiprocessor tests
for the VFS have been left in kernel.c commented for possible future
testing. Though, with the current modules, the VFS does pass the test.