Ported MB2BSP ELF loader, Makefile protections
I ported over the ELF loader from MB2BSP, I am pretty sure it is
working, the output seems correct, though I have not actually been able
to test it yet. For that I will need to get the main thread of a process
running. The Makefile has also been outfitted with a check such that if
$(ARC_ROOT) is empty, it will fill it with a . so that it doesn't
accidentally overwrite the root of the filesystem.