Fix ELF loader, module update, more config
The ELF program loader nows loads the program headers instead of the sections. The entire thing could be a little bit cleaner, but that's for the future. It does still need to map the headers in with proper permissions, but that should not be too difficult to do. Further config has been added to control the proportion of physical memory that is initialized into a buddy allocator vs. the amount that is left free for the freelist. The addition of a buddy allocator to the physical memory management stack will make DMA and process stack initialization easier.