terminal: increase memory and reorder putchar
The character memory has been increased from 4 KiB to 16 KiB. A check
has also been added to the draw function to ensure that it is not trying
to read outside of terminal memory. The write to the serial port has
also been moved from the end of the putchar function to the beginning.