-O1 flag, reorder tests
I have added the -O1 to the C flags. This is so that the compiler would
optimize its output just a little bit such that it does not repeat
itself constantly and move values to and from the stack unnecessarily.
The tests for the PMM have also been re-ordered to make the one that
should definitely fail happen towards the very end.