KPTI: add userspace section, enable interrupts
The interrupts have been enabled to allow the schedule to run. Further,
variables to track the beginning and end of the userspace section as
well as the userspace section itself have been added. A convencience
macro has also been added to more easily denote things that should go
into the userpace section.